Graft section parsing code into existing
This works in a test locally. The body text is parsed into sections, and each section scanned individually.
By doing that, we can keep track of which section a hit was in, and store it in the database

This looks like it works, in that it doesn't break anything and things turn up where I expect them to